Populate type attributes from type.__slots__
This change processes `type_obj.__slots__` to create descriptors of a new type `slot_descriptor` in `type_obj` in the type object initialization.

This change only handles correct population of such slot attributes in type objects, not get/set of them via instances of types with `__slots__`.

Implementing `slot_descriptor` 's get/set functions & prohibiting `__dict__` assignment will be done in a following diff.

To reuse `isidentifer` implementation, this change refactors its implementation and reuses its core logic in `strIsIdentifer` function in C++.

Note that `slot_descriptor` is not used as a data descriptor since it fails to be recognized as one by `typeIsDataDescriptor` in type-builtins.cpp and it's intentionally left out to isolate this change only to type dict generation.

Based on Facebook D20948400