Make types.coroutine() decorator work
In CPython 3.6's asyncio implementation many functions which could be transformed into coroutines are still implemented as generators. These are decorated with `@types.coroutine` to make them usable with the full coroutine implementation from PEP-492. This decorator is implemented in pure Python but to work it needs to be able to set the `CO_ITERABLE_COROUTINE` flag on `function.__code__`.

Without this feature the example from T65529554 fails with this error: P129012326.

Making `__code__` writable (T44845145) is probably out of scope for T65529554. This diff adds a quick work around for this problem by adding a specific Pyro native function to set the required flag.

Note: As mentioned in the diff below, I think the `CO_ITERABLE_COROUTINE` flag is just used as a safety guard by Python implementations to make sure generators are not used improperly with `await` expressions. So, another solution would be to ignore it entirely in Pyro. This would remove the need for this diff.

Note: Future versions of the asyncio implementation in CPython may no longer use  `@types.coroutine` internally. Maybe as early as 3.7 we might be able to skip this unless we need to support other libraries using this feature.

Based on Facebook D21033235