Move addInheritedSlots from runtime/type-builtins.cpp to ext/Object/typeobject.cpp
This change moves `addInheritedSlots` from `runtime` to the C-API layer, specifically its main user, `ext/Object/typeobject.cpp`.

This declutters `type-builtins.cpp` by moving the direct slot manipulation code of `addInheritedSlots` to C-API.

Note that `addInheritedSlots` is still called for `NativeProxy` types, but that logic should be moved to the C-API layer in the near future.

Based on Facebook D21232826