Add signal handlers to runtime
Following CPython, use `sigaction(2)` to set up signal handlers on the main thread. Forbid other threads from handling signals.

Store signal callbacks in the runtime, and maintain a bitfield to check which signals, if any, have been tripped. Modify the thread structure so we can mimic a stack overflow when we need to handle a signal.

Implement `PyErr_CheckSignals()` to notify the terminal when we have received a signal.

Based on Facebook D21137846