Add RawDataArray base class for LargeBytes, LargeStr, MutableBytes
The `LargeBytes`, `LargeStr` and `MutableBytes` share the same memory layout. Some operations (like searching for a sequence of bytes) make sense for any of the three classes so this introduces a common base class called `RawDataArray` so we only need to implement the algorithms once.

This removes the less useful `RawArrayBase` class which only had the `length()` method which is now duplicated into `RawTuple` and `RawDataArray`.

Based on Facebook D21229914