Revert D20948400 so that it can be landed with its following diffs
D20948400 breaks Distillery since its populated type attributes are unexpectedly interacting with the existing features of Python.

For example,  we found a class that populates a slot to be returned by `__str__`  as follows:

```
class C:
  __slots__ = "_value"

  def __str__(self):
    if hasattr(self, "_value"):
      return self._value
    else:
      self._value = ""
      return self._value
```
D20948400 populates a not-yet-working descriptor for `__slots__`, and this causes this class to return a descriptor value from `__str__` which breaks `C`'s client.

Note that this reverted diff is working correctly with its originally following diffs on the same stack (D20966390), so this will be grafted there to be submitted with them.

Based on Facebook D21363643