Reformat all py files
It seems there is a new version or different configuration for the python formatter in continuous integration. Reformat all .py files in `library/` that do not have `fmt: off` in one go, so we can keep the changes from other diffs.

While on it, fix some `str.__mod__` that meant to use tuples instead of just braces.

Based on Facebook D21381758