Build compile_commands.json by default
Summary: It's annoying having to remember how to set this whenever something goes wrong and I have to blow away my build directories. I don't think having this on is deterimental to anything, and it's supported by both Ninja and Makefile backends which I think are all we use.

Based on Facebook D21409063