Do not export pyro-internal type flags as type.__flags__
CPython exports the flag bitset of types as `__flags__` we used to export `RawType::flags()` as `__flags__` but that does not make much sense since our flags do not match up with cpython flags. This was leading to failures in the `copyreg` module which (optionally) reads and interprets those flags.

Stop providing a `__flags__` attribute on types.

Based on Facebook D21428547