Update caches based on queried state for monomorphic caching opcodes
This change fixes a bug that our runtime *always* upgrades a monomorphic caching opcode to the corresponding polymorphic one assuming that the cache state always remains monomorphic. This assumption is wrong since cache invalidation can revert the cache's state to anamorphic.

This change fixes the aforementioned problem by dynamically querying the state of a cache before updating the cache. This way, we can re-establish the lost invariant by invalidation at runtime. This change adds `icCurrentState()` function to query the state of a cache, and this function is separately defined from `icLookupMonomorphic()` not to burden it with any extra computation since `icCurrentState()` is only needed for cache update (slow path).

Based on Facebook D21358029