Have newLayout return sealed layouts
This changes `Runtime::newLayout` to return sealed layouts and require the caller to use `Runtime::layoutSetTupleOverflow`. This turns out to be more natural as it starts out with the simpler thing and most layouts for builtin types are actually sealed.

This also refactors `initType()` code to make it possible to move the `kSealSubtypeLayouts` handling into `Runtime::computeInitialLayout`.

Based on Facebook D21385371