Don't change top of stack in GET_ANEXT
Summary: `GET_ANEXT` was incorrectly overwriting the value at the top of the stack. As per https://docs.python.org/3/library/dis.html#opcode-GET_ANEXT the incoming top of the stack should be left unchanged and a new value added as a result of this opcode.

Based on Facebook D21081878