Implement slot_descriptor to read and write in-object attributes for __slots__
This change implements `slot_descriptor`'s `__get__` and `__set__` functions and registers the type as a data descriptor to mediate attribute access for attributes mentioned in `type.__slots__`.

Note that `slot_descriptor` is a simple wrapper of `instanceGetAttr` and `instanceSetAttr` except that it filters out attributes added by `__slots__`, but not explicitly set by the user yet. To achieve that `__slots__` attributes are set to `Unbound` initially and considered not exsiting.

Based on Facebook D21033233