Use initializeXXXTypes function instead of Builtins subclass
This changes the code to declare a `initializeXXXTypes` function in the headers instead of creating a subclass of `Builtins`.

This prepares for removal of `Builtins` and `ImmediateBuiltins` which avoids templates, a default (template) parameters and most importantly avoiding the  split into `initialize` and `postInitialize`.

(This diff is split into 2 for review purposes: part 1 handles the tricky cases like immediates, postInitialize functions or immediate layouts; part 2 translates the uniform rest with the help of the script in P130244668. I plan to merge both parts before landing)

Based on Facebook D21382013