Add more entries to _imp.extension_suffixes list
The `_imp.extension_suffixes` function provides a list of suffixes that importlib looks for looking for a C extension module.

CPython for example reports `['.cpython-37m-fb-007-x86_64.so', '.abi3.so', '.so']` with the effect that when import `xxx` it will search for `xxx.cpython-37m-fb-007-x86_64.so`, `xxx.abi3.so` and `xxx.so` and use whatever it can find first.

We would like to be able to produce pyro specific modules as well that cpython ignores, so we change the PyRo list to this now: `['.pyro.so', '.abi3.so', '.so']`.

While on it add some more entries and a comment to sysconfigdata.

Based on Facebook D21487776