Implement ctypes.memset with integer arguments
Summary: This implements the simple use-case of calling `ctypes.memset` with integer arguments. CPython allows for more argument types, but this is the minimal case needed by `multiprocessing`, and the `memset` special case in `CFuncPtr.__call__` will eventually be removed once we have a comprehensive argument marshalling strategy.

Based on Facebook D21525259