Replace Type::isSealed() with Type::hasMutableDict() to return type dict's mutability, not its layout
This change modifies Type::isSealed() to make it return true only when the given type's type dict is immutable.
Currently, Type::isSealed() returns true when its instance layout is sealed, which means the given type is modifiable whereas instances of the type cannnot be added with new attributes.

Based on Facebook D21484558