Store type.subclasses as a list instead of a dict
The only user-accessibly way to the subclass list is the `type.__subclassess` method which creates a copy of the whole list. This switches the internal representation to use a list instead of a dictionary. Add code to cleanup `None` entries into `type.__subclasses__` as that is iterating over the list anyway, so there is no change in complexity.

This will help in upcoming diffs that compute the type layout at a very late point of type initialization so we wouldn't have a layout id yet when adding to subclasses of a type.

Based on Facebook D21100450