Untrack objects after tp_dealloc even if the memory was not freed
This fixes the typical free-list pattern in C-API code where instead of freeing an object in the `tp_dealloc` callback you keep it around and re-use it later by calling `PyObject_Init` instead of allocating a new block of memory.

Context:
In PyRo for every extension object on the C/C++ heap we create a corresponding native proxy object on the managed heap. We also maintain a linked list of all extension objects because they contain a reference to the native proxy object which is a garbage collection root.

The tracking and native proxy setup is done in the `PyObject_Init` step. We used to untrack in the `PyObject_Free` / `PyObject_Del` functions, however that turned out to be tricky since some extension code does not actually free memory in the `tp_dealloc` slot but instead keeps it around in a free list for later re-use where it will call `PyObject_Init` on the memory block again to re-use it. However we should stop tracking the memory block in this case so a later `PyObject_Init` can work correctly.

Changes:
This changes `PyObject_Free` to reset the pointer in the native proxy to `None` if the memory of the extension object is freed. This allows us to detect after a `tp_dealloc` call whether the object memory was actually freed or is still around so we can untrack it as necessary.

Based on Facebook D21641235