Avoid normalization of tuples/StopIteration when a generator/coroutine returns
When using a coroutine we expect the exact value returned to propagate out via the `value` attribute of a `StopIteration` instance. However, because exception handling normalizes the `StopIteration` exception by calling [`createException()`](https://fburl.com/diffusion/kg5fq0b4) if the return value is a `tuple`, or `StopIteration` these will get slightly altered:
* A `tuple` like `(1, 2)` will get turned into `StopIteration(1, 2)` so the apparent value will be `1`.
* An explicit instance of `StopIteration(1)` will get turned into `StopIteration(1)` with an apparent value of `1` (not `StopIteration(1)`).

To counter this, we add a method for explicitly raising `StopIteration` instances: `Thread::raiseStopIterationWithValue()`. This explicitly instantiates a `StopIteration` exception containing the return value, which effectively turns the normalization into a no-op.

Handling the `StopIteration` case is a bit esoteric, but CPython does do this and it's tested for in `test_coroutines.py`.

This implementation is probably not the optimal end-state we want but it unblocks things for now. See T67598788.

Based on Facebook D21150608