Implement coroutine.close() + generator.close()
This feature was first introduced for generators in PEP-342. As PEP-342 became a stepping stone for full async/await coroutines in PEP-525, the semantics are mostly the same for both.

This diff implements both `coroutine.close()` and `generator.close()`, backed by the same implementation based on the [Python/pseudo-code in PEP-342](https://www.python.org/dev/peps/pep-0342/#new-generator-method-close).

Based on Facebook D21699616