Fix list.sort() to not compare items when their sort keys are equal
Summary: There is a bug in the current `list.sort()` implementation when sorting a list of **non-comparable** items with a provided `key` function, when the list contains items which generate duplicate keys. In that case, because we are sorting by a tuple of (key, item), the original item is then considered and crashes with `__lt__` not implemented.

Based on Facebook D21749070