Set initial layout's describedType to proper type
This change fixes a problem of leaving ill-formed layout objects in the heap by setting any created layout's described type to a proper type object.

With, `D20966390` typeInit() may create multiple layout objects, to create a finalized layout, In that process, it left some intermediate layouts that are not completely initialized. This partially initialized layouts caused `T67386267` and this diff fixes the issue by fully initializing it.

Based on Facebook D21699868