Implement coroutine.__await__()
As per the Python Data Model section 3.4.2 coroutine objects should have an `__await__()` method.

Rather than just returning `self` as you might expect, this returns a wrapper object around the `coroutine` instance which is iterable. This is needed because `coroutines` are explicitly not iterable according to [PEP-492](https://www.python.org/dev/peps/pep-0492/#differences-from-generators). The resulting `coroutine_wrapper` type in this diff is similar to the implementation in CPython.

Honestly, this seems excessive to me. I wouldn't be sorry if we just made `coroutines` iterable and returned `self` from `__await__()`.

Based on Facebook D21700916