Remove static from signal related runtime fields
None of the accessors are static and the code seems to work fine without `static`.

Without this change when we create 1 runtime instance and initialize signals a dangling reference will stay around in `signal_callbacks_` which will be picked up by the next runtime instance (if the next instance does not load the signals module and reinitializes the field). This appears to be the reason for the crashes some people get when running the tests sequentially.

Based on Facebook D21792694