Make GET_ANEXT behave like GET_AWAITABLE on result of calling __anext__()
After calling `__anext__()` on the target, `GET_ANEXT` should follow the same operation as `GET_AWAITABLE` on the same target. Specifically it must not attempt to call `__await__()` on an `async_generator` instance because that method doesn't exist.

This diff factors out and re-uses the common functionality from `GET_AWAITABLE` into `Interpreter::awaitableIter()`. This function is roughly analogous to CPython's `_PyCoro_GetAwaitableIter()`.

Based on Facebook D21083780