Special-case instance_proxy.__reduce__
This diff implements support for pickling an object's `__dict__`

Because we can't re-inflate an instance proxy, we must pickle it as a `dict`. This requires us to go around the usual machinery for object reduction, because the pickle library [checks](https://hlrz.com/source/xref/cpython-3.7/Lib/pickle.py?r=d561f848#629) that the object's class is equal to the serialized class if you use  `copyreg.__newobj__` as the object's "reinflater". Instead, we return `dict` as the construction method.

This is exercised by `lib2to3`, which pickles its grammar as a cache for future invocations.

Based on Facebook D21862706