Avoid duplicate typeslot initialization
`PyType_FromSpec` calls `typeNew` to create a new type and then initializes the type slots and as the last step inherits slots from super types. This means there is no need for `typeNew` to invoke the typeslot inheritance logic upfront, the results will be overwritten anyway. This changes the code to no longer perform type slot inheritance when using `typeNew` but still do so for `_builtins._type_init`.

While on it do some cleanups:
- Rename `addInheritedSlots` to `typeInheritSlots`.
- Add a `thread` parameter so we don't need to rely on `Thread::current()`.

Based on Facebook D21883956