Implement array.array.__{g,s}etitem_, __new__, __len__, extend, and append
Adds a list initializer to `array.array` by calling `extend` on the argument, which in turn calls `append` on each item of the iterable.

Also implements `array.array.__getitem__` and `array.array.__len__` to effectively test that the elements were properly set.

`array`s act mostly the same as `memoryview`s, but are backed by `MutableBytes` (and have a few different typecodes), therefore the logic and code for most of this functionality is taken from `memoryview.__getitem__` and `memoryview.__setitem__`.

Based on Facebook D19669615