Leave freeze_modules output files untouched when they would not change
This changes the `freeze_modules.py` to only write the output file if it does not exist yet or when the contents change. With this the file modification date will stay the same when the output did not change reducing time the build system spends on unnecessary rebuilding.

While on it also switch to a style where we write to a temporary file and then rename to the desired target filename. This helps having the operation being "atomic" and won't leave around half-written files in case of errors or aborts.

Based on Facebook D22021694