Always initialize parser->pos to 0 in parserInit
`_PyArg_Parser`'s are often declared statically and reused. This fixes a bug where the second time one was used, if there were positional-only args AND keywords, the parser would fail to parse.

This code was initially copied from CPython and a bug was introduced by a small refactoring.

Based on Facebook D22030095