Set prefix relative to executable; change builddir layout
This change the layout of the files in the build directory to match a typical unix project installation (with `<prefix>` being either `build/` or something like `/usr/local`). This also matches the "posix_prefix" scheme in `library/sysconfig.py`.
```
<prefix>/bin/python
<prefix>/include/pyro3.7     # for the C-API Headers
<prefix>/lib/pyro3.7         # for the standard library
```
We used to immitate a cpython build directory before by creating `Module/Setup` but adapting to the same scheme for installation and the build directory should simplify things and avoid confusion.
- Set sys.prefix to be the directory of the binary with `../lib/pyro3.7` appended.
- Adapt benchmark package and runner to new `python` binary path; Setting the `PYTHONHOME` environment variable in the benchmark runner is no longer necessary.

Based on Facebook D22095246