Update standard library file lists
We currently copy PyRos standard library into the `build/Lib` folder as part of the build process. However the result is not used yet except for the benchmark sent to the benchmark server. So unsurprisingly this list of files was incomplete and out-of-date now.

This tries to reduce friction and making it easier to keep the list up-to-date:
- Separate the list of standard library files from `CMakeLists.txt` and move it into `library/FILES` and `library/CPYTHON_FILES` with a simple 1 filename per line syntax.
- Add `util/update_library_files.sh` script to regenerate the lists with the current contents of `library/` and `third-party/cpython/Lib`.

Based on Facebook D22078678