Switch pyro to platform009
Now that we no longer rely on a system python being present we can switch from platform007 to platform009.

This removes the gcc+asan build as that is no longer working in platform009 (see https://fb.workplace.com/groups/asktp/permalink/2331098233853737/ ). We still have the clang+asan build so we should be good.

This also required some changes to the cmake build:
- Stop adding `-pthread` flag universally as it would confuse the cpython build when passing it in `CFLAGS` there. Instead just specify it for the pyro targets that actually need it.
- Pass flags as `CFLAGS` instead of `CFLAGS_NODIST` to cpython build, or we wouldn't end up using the `--platform=platform009` when building the extension modules via `setup.py` there. While on it also pass on `{CMAKE_C_FLAGS_${CMAKE_BUILD_MODE}}` which was previously missing.

Based on Facebook D22301367