Create python3 and python3.7 symlinks
Packaging for TP2 and buck requires that we we have a `python3` and a `python3.7` in the `bin` folder. This change the cmake build to create a `python3.7` binary with `python` and `python3` symlinks pointing to it.

While we could create symlinks when packaging for TP2, I think it is best to mirror the structure of an installed package in our build directory and adding it into the cmake build as well.

Based on Facebook D22420129