Handle deque.append and deque.appendleft over maxlen
When an item is added to a deque that increases its size to over maxlen, pop or popleft should be called depending on whether its append or appendleft.
Filed tasks for the implementation of setitem and reverse and commented out the Tests in the meantime. _maxlen and length no longer exist, instead we have maxlen and len(deque).

Based on Facebook D22283999