Simplify PyUnicode_DecodeLocale implementation
Summary: `PyUnicode_DecodeLocale`'s implementation was simplified betweeen 3.6 and 3.7. This updates the implementation to 3.7 (by mostly copy pasting the implementation and making use of the shared `_Py_DecodeLocaleEx` function in `fileutils`) to ensure that we call the same decoding logic, and always have throw the same exceptions.

Based on Facebook D22527075