Fix deque.popleft bug
Bug repro: Appending left results on the head pointing to the last index of the underlying tuple(which is desired) but a subsequent popleft results in the head incrementing beyond the capacity of the underlying tuple without wrapping around.

Fix: wrap around when head goes over capacity - 1.

Based on Facebook D22573424