Replace set lookup with CPython Implementation
The current implementation involves linear probing for set lookup and insertion. The CPython implementation uses a combination of a limited number of linear probes + linear congruential generator which adds an element of 'randomness' to the order of probing. The intention is that this will help reduce the number of collisions during insertions and lookups.

The performance improvement however is mostly a result in the reduction in the number of comparisons between keys. Previously, this was done every time the slot contained an item or tombstone but now this comparison is only made if the hash values match. This reduced the number of comparisons in the go benchmark from over 700k to only 182.

Additional performance information can be found in the perf-report of this diff

The set is grown whenever 40% of the set is occupied. This is a lower threshold than the CPython implementation which grows at 60% because it includes tombstones where as this is only actual items

Based on Facebook D22562305