Add throw operations to asynchronous generators
Much as we can send a value into a suspended async generator, we can raise an exception at the point where we did a generator-like yield. Essentially this is a generalization of the close operation which implicitly raises `GeneratorExit`.

Apart from throwing a parameterizable exception type/value, the implementation of `async_generator_athrow.send()` is subtly different from `async_generator_aclose.send()`. Particularly, if a `StopAsyncIteration` propagates out of the generator (indicating a `return`) this exception is passed on verbatim, whereas the close would turn this into a `StopIteration`. It's also okay for the throw to result in a generator-like yield, whereas in a close operation this raises a `RuntimeError` as it's indicative of a programming error.

I'm not sure why there needs to these differences, but I guess overall the close is a simplified and more constrained API for closing. Using the full throw for the same purposes gives the caller more control and information about what happened in the generator.

Based on Facebook D22876585