Implement builtins.open with bytes filenames and add Bytes::toCStr
Added a case for when path is a bytes object in `_os_open`.

`File::open` takes a `char*`, which means we need to copy our `Bytes` into a `NUL`-terminated `char*`, so this diff also adds a `toCStr` method to `Bytes` and `SmallBytes` (`LargeBytes` inherits its implementation from `DataArray::toCStr`).

Based on Facebook D22924029