Do not modify __context__ when an exception value is returned from a generator
When a generator returns a value it gets wrapped in a raised `StopIteation`. If the returned value happens to be an exception, this can incorrectly get caught up in the proces of exception chaining via `Thread::raiseStopIterationWithValue()` -> `Thread::raise()` -> `Thread::chainExceptionContext()`.

This diff fixes things by skipping that process if the value returned from a generator value is an exception. There may be other ways of solving this, but in this case I used the equivalent fix needed for CPython which had the same problem.

I came by this error in this CPython test:

https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/pyro/third-party/cpython/Lib/test/test_coroutines.py?commit=94b5a3e0f2693ef4a37badbad4332dab28459fc4&lines=1147-1161

And if you follow through the URL link in that test you'll find the fix for CPython.

Based on Facebook D23004049