Chain exception context across further generator calls in handler
Summary: When generators are suspended/resumed their caught-exception state needs to be preserved. This savable/restorable exception state is stored in a stack which grows as nested generators are started. As result, when raising an exception we must search through the stack of generator caught-exception states to see if there is an exception context to chain.

Based on Facebook D22994227