Avoid wrapping types with NativeProxy when possible
Previously, Pyro *unconditionally* wraps any extension type with `NativeProxy` type to reserve extra space to store `native` pointer pointing to a C-heap-allocated object.

This diff avoids such wrapping when possible: when a newly created type's instance size is 0 (`basicsize == itemsize == 0`), and the created type's base satisfizes the same condition, there's no need to allocate any `native` object, so thus we don't need to wrap such a type with `NativeProxy`. A direct effect of this is that we purely allocate an instance of such a type in Pyro's managed heap.

When an extension type is NOT wrapped in `NativeProxy`, its instance allocation should be done in the exactly same way that a managed type's instance is done. Therefore, the type's `tp_new` slot is populated with the same one for `managed_type.__new__`.

**Side note**

This change was motivated to support type creation with `metaclass` via C-API. Without this change, `subtype_of_type(new_type_name, bases, type_dict)` end up calling `PyType_GenericNew`, bypassing the type initialization logic in `type.__new__`.

Simply populating `type.__new__` in a subtype of `type` does not work either, since the rest of the system recognizes such a subtype as a `NativeProxy`, which will fail since `type.__new__` creates a type instance which is NOT a `NativeProxy`.

Therefore, the final solution is as proposed in this diff: letting a subtype of type inherit `type.__new__` and also labels such a `subtype` as a NOT `NativeProxy` type.  Note that subtypes of type are strictly used as a metaclass in Boost, so this usecase is well supported by this diff.

Based on Facebook D22941554