Rework Runtime constructor
Currently we have to make these choices before we can do anything in the runtime at all:
* Pick a maximum heap size
* Decide between C++ and assembly generated interpreter
* Initialize random state (so we can hash and intern strings)

This turns these 3 things into parameters of the `Runtime` class, so we stop doing ad-hoc queries for the `PYRO_CPP_INTERPRETER` environment variable and instead of things configurable by the runtime user.

Adds a struct to contain the random state and adds functions to initialize it with the OS secure random source or by a given fixed seed.

Based on Facebook D23000113