Rework builtins module initialization
Summary: This changes the builtin initialization code to update the cached instances in the runtime as soon as they are available. It also should initialize the `ctor` functions at type creation time. This ensures that the runtime is ready as soon as possible, so we can import other modules at the end of `builtins.py` in the next change.

Based on Facebook D23024945