Add type.__flags__
Note that this is different from the already extant `type.flags()` and
`Type::Flags` -- those are internal. This adds support for a property on `type`
that calculates CPython-specific flags on the fly based on information on the
type.

Also add `typeGetFlags` for extension types that set `tp_flags`.

Based on Facebook D23093512