Add PyRoAPI_FUNC
This replaces PyAPI_FUNC for functions declared in
`capi/cpython-func.h`. This (plus the following diff) allows us to generate
both declarations and definitions for these functions for an empty
`libpython.so`. If we used `PyAPI_FUNC`, it would also affect downstream users
of this header, like in `third-party/cpython/Modules`.

Based on Facebook D23255460