Reorder initialization code
Reorder initialization code in `builtins.py` so that no type is used anymore before it is defined with a `class` block.

This was previously working because `addModuleTypes` fills the module directories with half-initialized types before the module bodies start executing. This prepares for the next two changes that remove this early addition of the types to the namespace.

Note that there are 3 instances left where the type is already used inside the class body before the class body finished, which cannot be avoided by reordering:
* Construction of `exception.__cause__` and `exception.__context__`.
* `NoneType.__class__`

Based on Facebook D23163113