Use traceback object instances for tracebacks
Create instances of the traceback object to assign to the traceback value. This helps code that retrieves tracebacks and tries to assign them with `PyException_SetTraceback()` which checks that the parameter is a traceback object.

This is not a complete implementation of tracebacks, it still uses the same hack with a string-encoded traceback. It just gets assigned to the tracebacks frame slot now.

Based on Facebook D23409222