Fix PyErr_CheckSignals failing to run interrupt handler
There was a problem when a signal occurred and `PyErr_CheckSignals` was called without having the interpreter execute a function in between. In this case would enable "interrupt" mode on the thread but no function would be called yet. `handlePendingSignals` would then still have the thread in interrupt mode, and executing the signal handler function would fail because the thread interrupt situation would fire.

Fix this by having `Runtime::handlePendingSignals` clear the interrupted state of the thread.

This fixes crashes when the select module was polling and `CTRL+C` was pressed; this was the case in the django100 benchmark.

Based on Facebook D23521059