Check results in 2to3 benchmark and stop printing to stdout
Summary: Change 2to3 benchmark to stop printing to stdout (which among other things broke `diffrevs --run-benchmarks` and needed to be worked around in `_time_tool.py` because the output interfered with the measurement results. Instead it prints to a StringIO() object now and compares the result with expected outputs.

Based on Facebook D23618430