Store LayoutId on type
This diff adds another field to `RawType` to cache the `LayoutId` from the
`instanceLayout`. We'll use this in service of changing the inline caching
mechanism from using the address of the type to the `LayoutId` of the type's
`instanceLayout`.

Also do a drive-by cleanup of `RawType::sealAttributes`, since nobody
was using it.

Based on Facebook D23484373