Rework type layout creation
Refactor code: Put all the code that is related to creating a new layout in `typeInit` into a `typeComputeLayout` function and delay computation of the layout until the end of the function.

This also moves a number of runtime helpers (`computeInitialLayoutWithSlotAttributes`, `computeInitialLayout`, `computeInitialLayoutWithSlotAttributes`) that were only used by `typeInit` to `type-builtins.cpp`.

Based on Facebook D23522363