Remove Interpreter::callFunction with tuple argument
Summary: Remove `Interpreter::callFunction` function. This was a function calling variant where the parameters are packed in a tuple. However there was only a single caller of it and I'd rather not make it easy or encourage packing arguments into a tuple to perform a call.

Based on Facebook D23625688