Recalculate metaclass in type.__new__
`type.__new__` needs to recalculate the metaclass based on the bases list and if it changed, delegates to the metaclasses `__new__` function.

While on it also add a check whether a base is not a type and has `__mro_entries__` and raise a type error if so.

Based on Facebook D23617154