Add BuiltinFunction type
We used `Function::Entry` as the type for function trampolines and for the builtin functions. However they are different concepts and I plan to change one but not the other soon. So this introduces a new `BuiltinFunction` type and uses it in all code related to builtin functions.

Remove the `NativeMethodType` which existed for the same purpose but wasn't really used by anything but some test code...

Based on Facebook D23625620