Fix rebuild when library_files.txt or cpython_library_files.txt change
It seem that the approach of using `file(READ ...)` for `library_files.txt` and `cpython_library_files.txt` did not work great, because a rebuild of cmake would not be triggered when these two files changed. And I don't see any way to this up ritht now.

This changes the strategy to instead create a `library_files.cmake` and `cpython_library_files.cmake` file and use `include()` on them which does properly rebuild (and re-cmake) if the files changed.

Based on Facebook D23693840